backend/requirements.txt
Django==5.0
djangorestframework==3.15.0
drf-spectacular==0.27.1
flake8 >=3.9.2,<3.10
psycopg2==2.9.7
requests==2.25.1
django-cors-headers==4.1.0